+++ src/main/java/com/arangodb/internal/net/ExtendedHostResolver.java
@@ -20,8 +20,12 @@
 
 package com.arangodb.internal.net;
 
+import java.util.Arrays;
 import java.util.Collection;
 import java.util.List;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import com.arangodb.internal.util.HostUtils;
 
@@ -30,18 +34,23 @@ import com.arangodb.internal.util.HostUtils;
  *
  */
 public class ExtendedHostResolver implements HostResolver {
+	
+	private static final Logger LOGGER = LoggerFactory.getLogger(ExtendedHostResolver.class);
 
 	private static final long MAX_CACHE_TIME = 60 * 60 * 1000;
+	
 	private EndpointResolver resolver;
-	private final List<Host> hosts;
+	private HostSet hosts;
+	
 	private final Integer maxConnections;
 	private final ConnectionFactory connectionFactory;
+	
 	private long lastUpdate;
 
 	public ExtendedHostResolver(final List<Host> hosts, final Integer maxConnections,
 		final ConnectionFactory connectionFactory) {
 		super();
-		this.hosts = hosts;
+		this.hosts = new HostSet(hosts);
 		this.maxConnections = maxConnections;
 		this.connectionFactory = connectionFactory;
 		lastUpdate = 0;
@@ -53,23 +62,40 @@ public class ExtendedHostResolver implements HostResolver {
 	}
 
 	@Override
-	public List<Host> resolve(final boolean initial, final boolean closeConnections) {
+
+	public HostSet resolve(final boolean initial, final boolean closeConnections) {
+		
 		if (!initial && isExpired()) {
+			
 			lastUpdate = System.currentTimeMillis();
+			
 			final Collection<String> endpoints = resolver.resolve(closeConnections);
+			LOGGER.info("Resolve " + endpoints.size() + " Endpoints");
+			LOGGER.debug("Endpoints " + Arrays.deepToString(endpoints.toArray()));
+			
 			if (!endpoints.isEmpty()) {
 				hosts.clear();
 			}
+			
 			for (final String endpoint : endpoints) {
+				LOGGER.debug("Create HOST from " + endpoint);
+				
 				if (endpoint.matches(".*://.+:[0-9]+")) {
+					
 					final String[] s = endpoint.replaceAll(".*://", "").split(":");
 					if (s.length == 2) {
 						final HostDescription description = new HostDescription(s[0], Integer.valueOf(s[1]));
-						hosts.add(HostUtils.createHost(description, maxConnections, connectionFactory));
+						hosts.addHost(HostUtils.createHost(description, maxConnections, connectionFactory));
+					} else {
+						LOGGER.warn("Skip Endpoint (Missung Port)" + endpoint);
 					}
+					
+				} else {
+					LOGGER.warn("Skip Endpoint (Format)" + endpoint);
 				}
 			}
 		}
+		
 		return hosts;
 	}