+++ src/main/java/com/arangodb/internal/net/RandomHostHandler.java
@@ -63,7 +63,8 @@ public class RandomHostHandler implements HostHandler {
 	}
 
 	private Host getRandomHost(final boolean initial, final boolean closeConnections) {
-		final ArrayList<Host> hosts = new ArrayList<Host>(resolver.resolve(initial, closeConnections));
+		
+		final ArrayList<Host> hosts = new ArrayList<Host>(resolver.resolve(initial, closeConnections).getHostsList());
 		Collections.shuffle(hosts);
 		return hosts.get(0);
 	}
@@ -79,10 +80,8 @@ public class RandomHostHandler implements HostHandler {
 
 	@Override
 	public void close() throws IOException {
-		final List<Host> hosts = resolver.resolve(false, false);
-		for (final Host host : hosts) {
-			host.close();
-		}
+		final HostSet hosts = resolver.resolve(false, false);
+		hosts.close();
 	}
 
 	@Override