+++ src/main/java/com/arangodb/internal/velocypack/VPackDriverModule.java
@@ -23,12 +23,12 @@ package com.arangodb.internal.velocypack;
 import java.lang.reflect.Field;
 import java.util.Date;
 
-import com.arangodb.entity.ArangoDBVersion;
 import com.arangodb.entity.BaseDocument;
 import com.arangodb.entity.BaseEdgeDocument;
 import com.arangodb.entity.CollectionStatus;
 import com.arangodb.entity.CollectionType;
 import com.arangodb.entity.DocumentField;
+import com.arangodb.entity.License;
 import com.arangodb.entity.LogLevel;
 import com.arangodb.entity.Permissions;
 import com.arangodb.entity.QueryEntity;
@@ -37,8 +37,8 @@ import com.arangodb.entity.ReplicationFactor;
 import com.arangodb.entity.ViewType;
 import com.arangodb.entity.arangosearch.ArangoSearchProperties;
 import com.arangodb.entity.arangosearch.ArangoSearchPropertiesEntity;
-import com.arangodb.entity.arangosearch.ConsolidationType;
 import com.arangodb.entity.arangosearch.ConsolidationPolicy;
+import com.arangodb.entity.arangosearch.ConsolidationType;
 import com.arangodb.internal.velocystream.internal.AuthenticationRequest;
 import com.arangodb.model.TraversalOptions;
 import com.arangodb.model.arangosearch.ArangoSearchPropertiesOptions;
@@ -78,8 +78,7 @@ public class VPackDriverModule implements VPackModule, VPackParserModule {
 		context.registerSerializer(Permissions.class, VPackSerializers.PERMISSIONS);
 		context.registerSerializer(ReplicationFactor.class, VPackSerializers.REPLICATION_FACTOR);
 		context.registerSerializer(ViewType.class, VPackSerializers.VIEW_TYPE);
-		context.registerSerializer(ArangoSearchPropertiesOptions.class,
-			VPackSerializers.ARANGO_SEARCH_PROPERTIES_OPTIONS);
+		context.registerSerializer(ArangoSearchPropertiesOptions.class, VPackSerializers.ARANGO_SEARCH_PROPERTIES_OPTIONS);
 		context.registerSerializer(ArangoSearchProperties.class, VPackSerializers.ARANGO_SEARCH_PROPERTIES);
 		context.registerSerializer(ConsolidationType.class, VPackSerializers.CONSOLIDATE_TYPE);
 
@@ -90,14 +89,13 @@ public class VPackDriverModule implements VPackModule, VPackParserModule {
 		context.registerDeserializer(BaseEdgeDocument.class, VPackDeserializers.BASE_EDGE_DOCUMENT);
 		context.registerDeserializer(QueryEntity.PROPERTY_STARTED, Date.class, VPackDeserializers.DATE_STRING);
 		context.registerDeserializer(LogLevel.class, VPackDeserializers.LOG_LEVEL);
-		context.registerDeserializer(ArangoDBVersion.License.class, VPackDeserializers.LICENSE);
+		context.registerDeserializer(License.class, VPackDeserializers.LICENSE);
 		context.registerDeserializer(Permissions.class, VPackDeserializers.PERMISSIONS);
 		context.registerDeserializer(QueryExecutionState.class, VPackDeserializers.QUERY_EXECUTION_STATE);
 		context.registerDeserializer(ReplicationFactor.class, VPackDeserializers.REPLICATION_FACTOR);
 		context.registerDeserializer(ViewType.class, VPackDeserializers.VIEW_TYPE);
 		context.registerDeserializer(ArangoSearchProperties.class, VPackDeserializers.ARANGO_SEARCH_PROPERTIES);
-		context.registerDeserializer(ArangoSearchPropertiesEntity.class,
-			VPackDeserializers.ARANGO_SEARCH_PROPERTIES_ENTITY);
+		context.registerDeserializer(ArangoSearchPropertiesEntity.class, VPackDeserializers.ARANGO_SEARCH_PROPERTIES_ENTITY);
 		context.registerDeserializer(ConsolidationPolicy.class, VPackDeserializers.CONSOLIDATE);
 	}